replace baseFeeParams with alloy-eips-basefee-params

+++ crates/primitives/src/basefee.rs
@@ -30,7 +30,7 @@ pub fn calculate_next_block_base_fee(
     base_fee_params: crate::BaseFeeParams,
 ) -> u64 {
     // Calculate the target gas by dividing the gas limit by the elasticity multiplier.
-    let gas_target = gas_limit / base_fee_params.elasticity_multiplier;
+    let gas_target = gas_limit / base_fee_params.elasticity_multiplier as u64;
 
     match gas_used.cmp(&gas_target) {
         // If the gas used in the current block is equal to the gas target, the base fee remains the
